requirements_simple.txt
streamlit
google-generativeai
pillow
pypdf2
